move shared trace clearance helpers
