+++ lib/solvers/GroundedLoadPairSolver/GroundedLoadPairSolver.ts
@@ -5,7 +5,7 @@ import { BaseSolver } from "../BaseSolver"
 import { visualizeInputProblem } from "../LayoutPipelineSolver/visualizeInputProblem"
 import { getGroundedLoadPairs } from "./getGroundedLoadPairs"
 import { layoutGroundedLoadPair } from "./layoutGroundedLoadPair"
-import { offsetCollinearGroundedResistorLoads } from "../PackInnerPartitionsSolver/offsetCollinearConnections"
+import { offsetCollinearGroundedResistorLoads } from "../offsetCollinearConnections"
 
 export class GroundedLoadPairSolver extends BaseSolver {
   outputLayout: OutputLayout | null = null
